refactor: Decompose NativeDriver to opaque interop drivers
This change adds an abstraction layer over the interop details (such as
marshalling strings and handling error response codes) using the
typestate pattern to encapsulate the possible operations. This means
that all the interaction builders don't need to keep track of interop
concerns such as pact handles and interaction handles because those are
all encapsulated in the abstraction layer.

Note that HTTP pacts and message pacts are different in that multiple
HTTP interactions can be added to a pact before the mock server is
started and the file is written, whereas only one message interaction
can be configured/verified before the file is written. This creates a
slight asymmetry in the API for where the file writing method is defined,
but for good reason.

This should pave the way for adding specification v4 support.