fix: v4 message interactions weren't retaining property values
The v4 schema which is returned from the `Reify` call is different in v4
and so the deserialisation wasn't working, which meant a default message
was always sent for verification in consumer tests.

Instead it now passes the specification version all the way down to the
verification stage so we can deserialise the reified message properly
for verification. I'm not particularly happy with that as an approach,
but a proper typed version would be a breaking API change for something
which can be dealt with internally.