refactor: Overhaul the sample project
The old sample was confusing because it tried to do too much for one,
and also because the concepts of HTTP and message interactions were
split across multiple examples. There was also a separate Readme sample
which effectively just did the same thing as the canonical example.

Now there's one combined sample which demonstrates the most commonly
used features of both HTTP and message interactions, matchers and
provider states. It's presented in a much more realistic scenario of a
fulfilment component which responds to orders being created via pubsub
and then updates the order, so should be easier to intuitively
understand.